% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolve_ibd_dists.R
\name{convolve_ibd_dists}
\alias{convolve_ibd_dists}
\title{Convolve IBD distributions to obtain the distribution of the sum}
\usage{
convolve_ibd_dists(
  ...,
  point_mass_eps = 1e-09,
  number_of_gridpoints_exponent = 12
)
}
\arguments{
\item{...}{ibd dists.}

\item{number_of_gridpoints_exponent}{Default is 12.}
}
\value{
ibd_dist
}
\description{
Convolve IBD distributions to obtain the distribution of the sum
}
