% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_methods.R
\name{d}
\alias{d}
\title{Probability function for IBD Distribution Objects}
\usage{
d(x, ...)
}
\arguments{
\item{x}{An object of class \code{total_ibd_dist} or \code{segment_count_dist}.}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric value.
}
\description{
Returns the probability function of a total IBD or segment count distribution.
}
