% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_methods.R
\name{E}
\alias{E}
\title{Expectation for IBD Distribution Objects}
\usage{
E(x, ...)
}
\arguments{
\item{x}{An object of class \code{total_ibd_dist} or \code{segment_count_dist}.}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric value.
}
\description{
Computes the expectation (mean) of a total IBD or segment count distribution.
}
