% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_methods.R
\name{sd}
\alias{sd}
\title{Standard Deviation for IBD Distribution Objects}
\usage{
sd(x, ...)
}
\arguments{
\item{x}{An object of class \code{total_ibd_dist} or \code{segment_count_dist}.}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric value.
}
\description{
Computes the standard deviation of a total IBD or segment count distribution.
}
