% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_methods.R
\name{var}
\alias{var}
\title{Variance for IBD Distribution Objects}
\usage{
var(x, ...)
}
\arguments{
\item{x}{An object of class \code{total_ibd_dist} or \code{segment_count_dist}.}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric value.
}
\description{
Computes the variance of a total IBD or segment count distribution.
}
