% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibdsegments_package.R
\docType{package}
\name{ibdsegments-package}
\alias{ibdsegments}
\alias{ibdsegments-package}
\title{ibdsegments: Identity by Descent Probability in Pedigrees}
\description{
Identity by Descent (IBD) distributions in pedigrees. A Hidden Markov Model is used to compute identity coefficients, simulate IBD segments and to derive the distribution of total IBD sharing and segment count across chromosomes.
}
\author{
\strong{Maintainer}: Maarten Kruijver \email{maarten.kruijver@esr.cri.nz} (\href{https://orcid.org/0000-0002-6890-7632}{ORCID})

}
\keyword{internal}
