% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibdsegments_package.R
\docType{package}
\name{ibdsegments-package}
\alias{ibdsegments}
\alias{ibdsegments-package}
\title{ibdsegments: Probability of IBD and IBD Segments}
\description{
What the package does (one paragraph).

xx
}
\details{
TODO
}
\author{
\strong{Maintainer}: Maarten Kruijver \email{maarten.kruijver@esr.cri.nz} (\href{https://orcid.org/0000-0002-6890-7632}{ORCID})

}
\keyword{internal}
