% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inheritance_space.R
\name{inheritance_space}
\alias{inheritance_space}
\title{Inheritance space for pedigree}
\usage{
inheritance_space(
  pedigree,
  persons,
  coefficients = "kappa",
  exploit_symmetries = TRUE
)
}
\description{
The \code{inheritance_space} function determines the IBD vectors for a pedigree.
}
