% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_cibd.R
\name{r_cibd}
\alias{r_cibd}
\title{Random generation for IBD on a continuous genome}
\usage{
r_cibd(
  n,
  pedigree,
  persons = pedtools::leaves(pedigree),
  coefficients = "kappa",
  ibd_state = 1L,
  chromosome_length = 267.77
)
}
\arguments{
\item{n}{Number of observations}

\item{pedigree}{Pedigree in \code{\link[pedtools:ped]{pedtools::ped}} form.}

\item{persons}{Persons for which IBD is observed. Defaults to \href{pedigree}{\code{pedtools::leaves}}.}

\item{coefficients}{One of \code{"kappa"}, \code{"identity"} or \code{"detailed"}.}

\item{ibd_state}{Default is 1.}

\item{chromosome_length}{Default is 267.77 cM (an estimate of the length of chromosome 1).}
}
\description{
The \code{r_cibd}
}
\examples{
## Basic example: IBD along one chromosome for half siblings
L <- 300
r_hs <- r_cibd(n = 1e4, pedigree = pedtools::halfSibPed(), chromosome_length = 300)

# half sibs alternate between IBD (state 1) and not IBD (state 0)
head(r_hs$samples)

# the total_length and number of segments per sample are also returned
head(r_hs$stats)

## Comparing half siblings and grandparent-grandchild
r_gp <- r_cibd(n = 1e4, pedigree = pedtools::linearPed(2), persons = c(1, 5),
               chromosome_length = 300)

hist(r_gp$stats$total_length)
hist(r_hs$stats$total_length)
}
